% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bufbuf-data.R
\docType{data}
\name{bufbuf}
\alias{bufbuf}
\title{Common toad body temperatures and associated operative temperatures}
\format{
An object of class \code{data.frame} with 99 rows and 2 columns.
}
\source{
Wouter Beukema \email{wouter.beukema@gmail.com}
}
\usage{
data(bufbuf)
}
\description{
Active field body temperature data of Belgian common toads (\emph{Bufo bufo}) 
and associated operative temperature data from agar models, collected during 
nocturnal surveys in spring. Data are in degrees Celsius.
}
\keyword{datasets}
