% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-documentation.R
\docType{package}
\name{ectotemp-package}
\alias{ectotemp}
\alias{ectotemp-package}
\title{ectotemp: Quantitative Estimates of Small Ectotherm Temperature Regulation Effectiveness}
\description{
Easy and rapid quantitative estimation of small terrestrial ectotherm temperature regulation effectiveness in R. ectotemp is built on classical formulas that evaluate temperature regulation by means of various indices, inaugurated by Hertz et al. (1993) <doi: 10.1086/285573>. Options for bootstrapping and permutation testing are included to test hypotheses about divergence between organisms, species or populations.
}
\references{
Blouin-Demers, G., & Weatherhead, P. J. (2001). Thermal ecology of black rat
snakes (\emph{Elaphe obsoleta}) in a thermally challenging environment.
Ecology, 82(11), 3025-3043. \cr
Blouin-Demers, G., & Nadeau, P. (2005). The cost-benefit model of
thermoregulation does not predict lizard thermoregulatory behavior. Ecology,
86(3), 560-566. \cr
Christian, K. A., & Weavers, B. W. (1996). Thermoregulation of monitor
lizards in Australia: an evaluation of methods in thermal biology.
Ecological monographs, 66(2), 139-157. \cr
Hertz, P. E., Huey, R. B., & Stevenson, R. D. (1993). Evaluating temperature
regulation by field-active ectotherms: the fallacy of the inappropriate
question. The American Naturalist, 142(5), 796-818. \cr
Wills, C. A., & Beaupre, S. J. (2000). An application of randomization for
detecting evidence of thermoregulation in timber rattlesnakes (\emph{Crotalus
horridus}) from northwest Arkansas. Physiological and Biochemical Zoology,
73(3), 325-334.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/wouterbeukema/ectotemp}
}
}
