% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ichalp-data.R
\docType{data}
\name{ichalp}
\alias{ichalp}
\title{Alpine newt body temperatures and associated operative temperatures}
\format{An object of class \code{data.frame} with 99 rows and 2 columns.}
\source{
Wouter Beukema \email{wouter.beukema@gmail.com}
}
\usage{
data(ichalp)
}
\description{
Active field body temperature data of Belgian alpine newts 
(\emph{Ichthyosaura alpestris}) and associated operative temperature data 
from agar models, collected during nocturnal surveys in spring. Data are in 
degrees Celsius.
}
\keyword{datasets}
